Feature/berrielocal 26

+++ src/main/kotlin/ru/vsu/cs/berrielocal/dto/product/ProductResponse.kt
@@ -5,8 +5,8 @@ data class ProductResponse(
     val name: String? = null,
     val imageUrl: String? = null,
     val cost: Double? = null,
-    val maxSize: Double? = null,
-    val minSize: Double? = null,
+    val maxSize: Long? = null,
+    val minSize: Long? = null,
     val units: String? = null,
     var shopId: Long? = null
 )
